requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pg8000==1.30.3
python-dotenv==1.0.0